cannon: Autodetect VM type from state in witness command
